+++ JamGame/Assets/ScriptableObjects/Employee Configs/Quirks/Glutton.asset
@@ -22,3 +22,7 @@ MonoBehaviour:
     DecreaseSpeed: 0.5
     OverrideSatisfaction: 0
     OverrideSatisfactionValue: 0
+  effects: []
+  references:
+    version: 2
+    RefIds: []
