+++ JamGame/Assets/Scripts/Employee/Buff.cs
@@ -1,7 +1,5 @@
-using System;
 using System.Collections.Generic;
 using Common;
-using Employee.Needs;
 using Level.GlobalTime;
 using Sirenix.OdinInspector;
 using UnityEngine;
@@ -21,53 +19,5 @@ namespace Employee
         [SerializeReference]
         private List<IEffect> effects = new();
         public IEnumerable<IEffect> Effects => effects;
-    }
-
-    public interface IEffectExecutor { }
-
-    public interface IEffectExecutor<E> : IEffectExecutor
-        where E : class, IEffect
-    {
-        public void RegisterEffect(E effect);
-        public void UnregisterEffect(E effect);
-    }
-
-    [HideReferenceObjectPicker]
-    public interface IEffect { }
-
-    [Serializable]
-    public class StressEffect : IEffect
-    {
-        [SerializeField]
-        [FoldoutGroup("Stress effect")]
-        private float increaseMultiplier;
-        public float IncreaseMultiplier => increaseMultiplier;
-    }
-
-    [Serializable]
-    public class NeedModifierEffect : IEffect
-    {
-        [SerializeField]
-        [FoldoutGroup("NeedModifier effect")]
-        private List<Need.NeedProperties> needModifiers = new();
-        public IEnumerable<Need.NeedProperties> NeedModifiers => needModifiers;
-    }
-
-    [Serializable]
-    public class ControllerEffect : IEffect
-    {
-        [SerializeField]
-        [FoldoutGroup("Controller effect")]
-        private float speedMultiplier;
-        public float SpeedMultiplier => speedMultiplier;
-    }
-
-    [Serializable]
-    public class EarnedMoneyEffect : IEffect
-    {
-        [SerializeField]
-        [FoldoutGroup("Earned money effect")]
-        private float multiplier;
-        public float Multiplier => multiplier;
     }
 }
