+++ JamGame/Assets/Scripts/Employee/Effect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f38ce3bd60e5be6418dec027eab37e73
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
